pyproject.toml
[project]
name = "licence-plate-detection" # Edit this and remove comment
version = "1.0.0"

requires-python = ">=3.13"
dependencies = [
    # "common-code[test] @ git+https://github.com/swiss-ai-center/common-code.git@main",
    "common-code[test] @ file:///Users/simonbraillard/development/pycharm_workspace/SwissAiCenter/common-code", # TODO: To delete
    "ultralytics==8.4.14",
    "pillow==12.1.1",
    "numpy==2.4.2"
]


[tool.pytest.ini_options]
pythonpath = [".", "src"]
addopts = "--cov-config=.coveragerc --cov-report xml --cov-report term-missing --cov=./src"
